+++ src/cli/index.ts
@@ -1,7 +1,10 @@
 import { createCommand } from 'commander';
 
-import { collect, visualize } from './commands';
+import { collect, visualize, sync } from './commands';
 
 export function createCli() {
-  return createCommand().addCommand(collect()).addCommand(visualize());
+  return createCommand()
+    .addCommand(collect())
+    .addCommand(visualize())
+    .addCommand(sync());
 }